commit 1

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -14,7 +14,7 @@ int main(void)
     int i;
     for (i = 1; i<= 98; i++)
     {
-        s = t0 + t1;
+        s = (unsigned long int)t0 + (unsigned long int)t1;
         if (i != 98)
         {
 			printf("%lu, ", s);
